examples/edna-py/edna-langchain/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "edna_langchain"
version = "0.1.0"
description = "Engagement Development Nudge Agent with LangChain"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-openai>=0.1.0",
    "faiss-cpu>=1.7.4",
    "tabulate>=0.9.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["edna*"]